+++ spec/philiprehberger/inflector_spec.rb
@@ -170,4 +170,168 @@ RSpec.describe Philiprehberger::Inflector do
       expect(described_class.titleize('UserAccount')).to eq('User Account')
     end
   end
+
+  describe '.ordinalize' do
+    it 'converts 1 to 1st' do
+      expect(described_class.ordinalize(1)).to eq('1st')
+    end
+
+    it 'converts 2 to 2nd' do
+      expect(described_class.ordinalize(2)).to eq('2nd')
+    end
+
+    it 'converts 3 to 3rd' do
+      expect(described_class.ordinalize(3)).to eq('3rd')
+    end
+
+    it 'converts 4 to 4th' do
+      expect(described_class.ordinalize(4)).to eq('4th')
+    end
+
+    it 'converts 11 to 11th' do
+      expect(described_class.ordinalize(11)).to eq('11th')
+    end
+
+    it 'converts 12 to 12th' do
+      expect(described_class.ordinalize(12)).to eq('12th')
+    end
+
+    it 'converts 13 to 13th' do
+      expect(described_class.ordinalize(13)).to eq('13th')
+    end
+
+    it 'converts 21 to 21st' do
+      expect(described_class.ordinalize(21)).to eq('21st')
+    end
+
+    it 'converts 22 to 22nd' do
+      expect(described_class.ordinalize(22)).to eq('22nd')
+    end
+
+    it 'converts 23 to 23rd' do
+      expect(described_class.ordinalize(23)).to eq('23rd')
+    end
+
+    it 'converts 100 to 100th' do
+      expect(described_class.ordinalize(100)).to eq('100th')
+    end
+
+    it 'converts 101 to 101st' do
+      expect(described_class.ordinalize(101)).to eq('101st')
+    end
+
+    it 'converts 111 to 111th' do
+      expect(described_class.ordinalize(111)).to eq('111th')
+    end
+
+    it 'converts 112 to 112th' do
+      expect(described_class.ordinalize(112)).to eq('112th')
+    end
+
+    it 'converts 113 to 113th' do
+      expect(described_class.ordinalize(113)).to eq('113th')
+    end
+
+    it 'handles 0' do
+      expect(described_class.ordinalize(0)).to eq('0th')
+    end
+
+    it 'handles negative numbers' do
+      expect(described_class.ordinalize(-1)).to eq('-1st')
+      expect(described_class.ordinalize(-11)).to eq('-11th')
+      expect(described_class.ordinalize(-21)).to eq('-21st')
+    end
+  end
+
+  describe '.dasherize' do
+    it 'converts underscored string to dashed' do
+      expect(described_class.dasherize('some_thing')).to eq('some-thing')
+    end
+
+    it 'handles already dashed strings' do
+      expect(described_class.dasherize('some-thing')).to eq('some-thing')
+    end
+
+    it 'handles strings without underscores' do
+      expect(described_class.dasherize('something')).to eq('something')
+    end
+
+    it 'handles multiple underscores' do
+      expect(described_class.dasherize('one_two_three')).to eq('one-two-three')
+    end
+  end
+
+  describe '.demodulize' do
+    it 'removes module namespace' do
+      expect(described_class.demodulize('Admin::User')).to eq('User')
+    end
+
+    it 'handles deeply nested namespaces' do
+      expect(described_class.demodulize('App::Admin::User')).to eq('User')
+    end
+
+    it 'returns the string when no namespace' do
+      expect(described_class.demodulize('User')).to eq('User')
+    end
+
+    it 'handles empty string' do
+      expect(described_class.demodulize('')).to eq('')
+    end
+  end
+
+  describe '.add_plural_rule' do
+    after do
+      described_class.custom_plurals.shift
+    end
+
+    it 'adds a custom pluralization rule with regex' do
+      described_class.add_plural_rule(/foo$/i, 'fooze')
+      expect(described_class.pluralize('foo')).to eq('fooze')
+    end
+
+    it 'adds a custom pluralization rule with string' do
+      described_class.add_plural_rule('bar$', 'barz')
+      expect(described_class.pluralize('bar')).to eq('barz')
+    end
+
+    it 'gives priority to custom rules' do
+      described_class.add_plural_rule(/^cat$/i, 'kittens')
+      expect(described_class.pluralize('cat')).to eq('kittens')
+    end
+  end
+
+  describe '.add_singular_rule' do
+    after do
+      described_class.custom_singulars.shift
+    end
+
+    it 'adds a custom singularization rule with regex' do
+      described_class.add_singular_rule(/fooze$/i, 'foo')
+      expect(described_class.singularize('fooze')).to eq('foo')
+    end
+
+    it 'adds a custom singularization rule with string' do
+      described_class.add_singular_rule('barz$', 'bar')
+      expect(described_class.singularize('barz')).to eq('bar')
+    end
+  end
+
+  describe '.add_uncountable' do
+    after do
+      described_class.custom_uncountables.pop
+    end
+
+    it 'adds a custom uncountable word' do
+      described_class.add_uncountable('funky')
+      expect(described_class.pluralize('funky')).to eq('funky')
+      expect(described_class.singularize('funky')).to eq('funky')
+    end
+
+    it 'adds multiple uncountable words' do
+      described_class.add_uncountable('alpha', 'beta')
+      expect(described_class.pluralize('alpha')).to eq('alpha')
+      expect(described_class.pluralize('beta')).to eq('beta')
+      described_class.custom_uncountables.pop
+    end
+  end
 end